add search logic
